backend/pyproject.toml
[project]
name = "guidelines-vs-traditional-prompt-backend"
version = "0.1.0"
description = "FastAPI Backend for Parlant Guidelines vs Traditional Prompt Comparison"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]
